requirements.txt
django==3.2.13
djangorestframework==3.13.1
drf-spectacular==0.22.0
coverage==6.3.2
pre-commit==2.18.1
djangorestframework-simplejw==5.1.0
requests==2.27.1